Tests using Prometheus
