+++ metrics/cpuMetrics.ts
@@ -1,9 +1,12 @@
 import Metrics from './Metrics';
 
+console.log();
+
 let queryString = 'camino_resource_tracker_cpu_usage{namespace="santi"}';
 let metrics = new Metrics();
 metrics.execPrometheus(queryString);
 
 process.on("disconnect", () => {
-    metrics.saveJSONData("cpuMetrics.json");
+    metrics.saveJSONData(`cpuMetrics${process.env.caseTest}.json`);
+    metrics.clearMetrics();
 })
